+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 6.26.0 (December 1, 2021)
 ### New Features
 - <a href="https://github.com/openscope/openscope/issues/1847" target="_blank">#1847</a> - Allow user to change the wind
+- <a href="https://github.com/openscope/openscope/issues/1848" target="_blank">#1848</a> - Allow user to toggle individual video maps on/off
 
 ### Bugfixes